+++ python/hgraph/__init__.py
@@ -21,7 +21,7 @@ from ._types import (TS, TSS, TSD, TSL, TSB, Size, TimeSeriesSchema, CONTEXT, RE
 from ._compat import (CmpResult, DivideByZero, exception_time_series, try_except,
                       TryExceptResult, TryExceptTsdMapResult, NodeError,
                       OperatorWiringNodeClass, BoolResult, CompoundScalar, JSON, TimeSeriesReference,
-                      NodeException, accumulate, average, center_of_mass_to_alpha, span_to_alpha,
+                      NodeException, accumulate, average,
                       to_json_builder, from_json_builder, register_json_datetime_format)
 from ._wiring import filter_by
 from ._wiring import convert